+++ lib/nnq/routing/surveyor.rb
@@ -24,7 +24,6 @@ module NNQ
       def initialize(engine)
         @engine     = engine
         @queues     = {} # conn => Async::LimitedQueue
-        @pump_tasks = {} # conn => Async::Task
         @recv_queue = Async::Queue.new
         @current_id = nil
         @mutex      = Mutex.new
@@ -85,21 +84,14 @@ module NNQ
 
 
       def connection_added(conn)
-        queue             = Async::LimitedQueue.new(@engine.options.send_hwm)
-        @queues[conn]     = queue
-        @pump_tasks[conn] = spawn_pump(conn, queue)
+        queue         = Async::LimitedQueue.new(@engine.options.send_hwm)
+        @queues[conn] = queue
+        spawn_pump(conn, queue)
       end
 
 
       def connection_removed(conn)
         @queues.delete(conn)
-        task = @pump_tasks.delete(conn)
-
-        return unless task
-        return if task == Async::Task.current
-
-        task.stop
-      rescue IOError, Errno::EPIPE
       end
 
 
@@ -109,8 +101,6 @@ module NNQ
 
 
       def close
-        @pump_tasks.each_value(&:stop)
-        @pump_tasks.clear
         @queues.clear
         @recv_queue.enqueue(nil)
       end